- shtimi i autorizimeve me ane te vlerave ne token, shtimi i fshirjeve te niveleve me te ulta
- jane shtuar autorizimet neper secilen metod bazuar nga vlerat e marra nga tokeni ne Boards, Lists, Tasks, Invite, Members dhe StarredBoard controllerat (si dhe verifikime tjera)
- jane shtuar fshirje ne metodat delete qe te fshihen eleementet e niveleve me te ulta psh kur fshihet bordi te fshihen listat e atij bordi
- eshte larguar atributi starred dhe eshte shtuar metoda e cila mundeson te marr bordet e bera starr nga nje user ne nje workspace

+++ backend/Controllers/BoardController.cs
@@ -4,6 +4,7 @@ using backend.DTOs.Board.Output;
 using backend.DTOs.Workspace;
 using backend.Interfaces;
 using backend.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace backend.Controllers
@@ -15,16 +16,21 @@ namespace backend.Controllers
         private readonly IBoardRepository _boardRepo;
         private readonly IWorkspaceRepository _workspaceRepo;
         private readonly IBackgroundRepository _backgroundRepo;
+        private readonly IMembersRepository _membersRepo;
+        private readonly IUserRepository _userRepo;
         private readonly IMapper _mapper;
 
-        public BoardController(IBoardRepository boardRepo,IWorkspaceRepository workspaceRepo, IBackgroundRepository backgroundRepo, IMapper mapper)
+        public BoardController(IBoardRepository boardRepo,IWorkspaceRepository workspaceRepo, IBackgroundRepository backgroundRepo, IMembersRepository membersRepo, IUserRepository userRepo, IMapper mapper)
         {
             _boardRepo = boardRepo;
             _workspaceRepo = workspaceRepo;
             _backgroundRepo = backgroundRepo;
+            _membersRepo = membersRepo;
+            _userRepo = userRepo;
             _mapper = mapper;
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [Authorize(Policy = "AdminOnly")]
         [HttpGet(template:"GetAllBoards")]
         public async Task<IActionResult> GetAllBoards()
         {
@@ -44,52 +50,72 @@ namespace backend.Controllers
                 return StatusCode(500, "Internal Server Error!"+e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetBoardsByWorkspaceId")]
         public async Task<IActionResult> GetBoardsByWorkspaceId(int workspaceId)
         {
             try
             {
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var isMember = await _membersRepo.IsAMember(userId, workspaceId);
+                
+                
                 if (!await _workspaceRepo.WorkspaceExists(workspaceId))
                 {
                     return NotFound("Workspace Not Found!");
                 }
-                
-                var boards = await _boardRepo.GetBoardsByWorkspaceIdAsync(workspaceId);
-                
-                if (boards.Count == 0)
+
+                if (isMember || userTokenRole == "Admin")
                 {
-                    return  Ok(new List<BoardDto>()); // Return an empty list if no boards are found
+
+                    var boards = await _boardRepo.GetBoardsByWorkspaceIdAsync(workspaceId);
+
+                    if (boards.Count == 0)
+                    {
+                        return Ok(new List<BoardDto>()); // Return an empty list if no boards are found
+                    }
+
+                    var boardDtos = _mapper.Map<IEnumerable<BoardDto>>(boards);
+                    return Ok(boardDtos);
                 }
 
-                var boardDtos = _mapper.Map<IEnumerable<BoardDto>>(boards);
-                return Ok(boardDtos);
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error!"+e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetBoardByID")]
         public async Task<IActionResult> GetBoardById(int id)
         {
             try
             {
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
                 var board = await _boardRepo.GetBoardByIdAsync(id);
 
                 if (board == null)
                     return NotFound("Board Not Found!");
 
-                var boardDto = _mapper.Map<BoardDto>(board);
-                return Ok(boardDto);
+                var workspaceId = board.WorkspaceId;
+                var isMember = await _membersRepo.IsAMember(userId, workspaceId);
+
+                if (isMember || userTokenRole == "Admin")
+                {
+                    var boardDto = _mapper.Map<BoardDto>(board);
+                    return Ok(boardDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
-                return StatusCode(500, "Internal Server Error!"+e.Message);
+                return StatusCode(500, "Internal Server Error!" + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpPost("CreateBoard")]
         public async Task<IActionResult> CreateBoard(CreateBoardDto boardDto)
         {
@@ -108,18 +134,26 @@ namespace backend.Controllers
 
             try
             {
-                var boardModel = _mapper.Map<Board>(boardDto);
-                await _boardRepo.CreateBoardAsync(boardModel);
-                return CreatedAtAction(nameof(GetBoardById), new { id = boardModel.BoardId }, _mapper.Map<BoardDto>(boardModel));
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var isMember = await _membersRepo.IsAMember(userId, boardDto.WorkspaceId);
+                if (isMember || userTokenRole == "Admin")
+                {
+                    var boardModel = _mapper.Map<Board>(boardDto);
+                    await _boardRepo.CreateBoardAsync(boardModel);
+                    return CreatedAtAction(nameof(GetBoardById), new { id = boardModel.BoardId },
+                        _mapper.Map<BoardDto>(boardModel));
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
-                return StatusCode(500, "Internal Server Error!"+e.Message);
+                return StatusCode(500, "Internal Server Error!" + e.Message);
             }
         }
 
-        [HttpPut]
-        [Route("UpdateBoard")]
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPut("UpdateBoard")]
         public async Task<IActionResult> UpdateBoard(UpdateBoardRequestDto updateDto)
         {
             if (!ModelState.IsValid)
@@ -129,22 +163,30 @@ namespace backend.Controllers
 
             try
             {
-                var board = await _boardRepo.UpdateBoardAsync(updateDto);
-    
-                if (board == null)
-                    return NotFound("Board Not Found!");
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var isMember = await _membersRepo.IsAMember(userId, updateDto.WorkspaceId);
 
-                var boardDto = _mapper.Map<BoardDto>(board);
-                return Ok(boardDto);
+                if (isMember || userTokenRole == "Admin")
+                {
+                    var board = await _boardRepo.UpdateBoardAsync(updateDto);
+
+                    if (board == null)
+                        return NotFound("Board Not Found!");
+
+                    var boardDto = _mapper.Map<BoardDto>(board);
+                    return Ok(boardDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error!"+e.Message);
             }
         }
-
-        [HttpDelete]
-        [Route("DeleteBoardByID")]
+        
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpDelete("DeleteBoardByID")]
         public async Task<IActionResult> DeleteBoard([FromQuery] BoardIdDto boardIdDto)
         {
             if  (!ModelState.IsValid)
@@ -152,19 +194,33 @@ namespace backend.Controllers
 
             try
             {
-                var boardModel = await _boardRepo.DeleteBoardAsync(boardIdDto.BoardId);
+                
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
 
-                if (boardModel == null)
-                    return NotFound("Board Does Not Exist!");
+                var board = await _boardRepo.GetBoardByIdAsync(boardIdDto.BoardId);
 
-                return Ok("Board Deleted!");
+                if (board == null) return NotFound("Board does not exists!");
+
+                var workspaceId = board.WorkspaceId;    
+                
+                //vetem owner mundet me delete board
+                var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, workspaceId);
+
+                if (ownsWorkspace || userTokenRole == "Admin")
+                {
+                    var boardModel = await _boardRepo.DeleteBoardAsync(boardIdDto.BoardId);
+                    return Ok("Board Deleted!");
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error!"+e.Message);
             }
         }
-
+        
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpDelete(template:"DeleteBoardsByWorkpaceID")]
         public async Task<IActionResult> DeleteByWorkspace(WorkspaceIdDto workspaceIdDto)
         {
@@ -178,79 +234,123 @@ namespace backend.Controllers
 
             try
             {
-                var boardModel = await _boardRepo.DeleteBoardsByWorkspaceIdAsync(workspaceIdDto.WorkspaceId);
-
-                if (boardModel.Count == 0)
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, workspaceIdDto.WorkspaceId);
+                //vetem owner mundet me delete board
+                if (ownsWorkspace || userTokenRole == "Admin")
                 {
-                    return NotFound("Boards Not Founded!");
-                }
+                    var boardModel = await _boardRepo.DeleteBoardsByWorkspaceIdAsync(workspaceIdDto.WorkspaceId);
 
-                return Ok("Boards Deleted!");
+                    if (boardModel.Count == 0)
+                    {
+                        return NotFound("Boards Not Founded!");
+                    }
+
+                    return Ok("Boards Deleted!");
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error!"+e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpPost("Close")]
         public async Task<IActionResult> CloseBoard(CloseBoardRequestDto dto)
         {
             try
             {
-                var result = await _boardRepo.CloseBoardAsync(dto.BoardId, dto.userId);
-                if (!result)
-                {
-                    return NotFound();
-                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var board = await _boardRepo.GetBoardByIdAsync(dto.BoardId);
 
-                return Ok("Board closed");
+                if (board == null) return NotFound("Board does not exists!");
+
+                var workspaceId = board.WorkspaceId;
+
+                //vetem owner mundet me bo close
+                var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, workspaceId);
+                if (ownsWorkspace || userTokenRole == "Admin")
+                {
+                    var result = await _boardRepo.CloseBoardAsync(dto.BoardId, dto.userId);
+                    if (!result)
+                    {
+                        return NotFound();
+                    }
+
+                    return Ok("Board closed");
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error!"+e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpPost("Reopen")]
         public async Task<IActionResult> ReopenBoard(CloseBoardRequestDto dto)
         {
             try
             {
-                var result = await _boardRepo.ReopenBoardAsync(dto.BoardId, dto.userId);
-                if (!result)
-                {
-                    return NotFound();
-                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var board = await _boardRepo.GetBoardByIdAsync(dto.BoardId);
 
-                return Ok("Board reopened");
+                if (board == null) return NotFound("Board does not exists!");
+
+                var workspaceId = board.WorkspaceId;
+
+                //vetem owner mundet me reopen
+                var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, workspaceId);
+                if (ownsWorkspace || userTokenRole == "Admin")
+                {
+                    var result = await _boardRepo.ReopenBoardAsync(dto.BoardId, dto.userId);
+                    if (!result)
+                    {
+                        return NotFound();
+                    }
+
+                    return Ok("Board reopened");
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
-                return StatusCode(500, "Internal Server Error!"+e.Message);
+                return StatusCode(500, "Internal Server Error!" + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetClosedBoards")]
         public async Task<IActionResult> GetClosedBoards(int workspaceId)
         {
             try
             {
-                var closedBoards = await _boardRepo.GetClosedBoardsAsync(workspaceId);
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var isMember = await _membersRepo.IsAMember(userId, workspaceId);
 
-                if (closedBoards.Count == 0)
+                if (isMember || userTokenRole == "Admin")
                 {
-                    return  Ok(new List<BoardDto>()); 
-                }
-                
-                
-                var closedBoardsDto = _mapper.Map<IEnumerable<BoardDto>>(closedBoards);
+                    var closedBoards = await _boardRepo.GetClosedBoardsAsync(workspaceId);
 
-                return Ok(closedBoardsDto);
+                    if (closedBoards.Count == 0)
+                    {
+                        return Ok(new List<BoardDto>());
+                    }
+
+
+                    var closedBoardsDto = _mapper.Map<IEnumerable<BoardDto>>(closedBoards);
+
+                    return Ok(closedBoardsDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
-                return StatusCode(500, "Internal Server Error!"+e.Message);
+                return StatusCode(500, "Internal Server Error!" + e.Message);
             }
         }
