+++ backend/Controllers/InviteController.cs
@@ -3,6 +3,7 @@ using backend.DTOs.Invite.Input;
 using backend.DTOs.Invite.Output;
 using backend.Interfaces;
 using backend.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace backend.Controllers
@@ -25,7 +26,8 @@ namespace backend.Controllers
             _mapper = mapper;
         }
 
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [Authorize(Policy = "AdminOnly")]
         [HttpGet(template: "GetAllInvites")]
         public async Task<IActionResult> GetAllInvites()
         {
@@ -43,7 +45,8 @@ namespace backend.Controllers
                 return StatusCode(500, "Internal Server Error " + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [Authorize(Policy = "AdminOnly")]
         [HttpGet("GetInviteById")]
         public async Task<IActionResult> GetInviteById(int id)
         {
@@ -51,8 +54,7 @@ namespace backend.Controllers
             {
                 var invite = await _inviteRepo.GetInviteByIdAsync(id);
                 if (invite == null) return NotFound("Invite Not Found!");
-               
-
+          
                 var inviteDto = _mapper.Map<InviteDtoOut>(invite);
                 return Ok(inviteDto);
             }
@@ -61,36 +63,57 @@ namespace backend.Controllers
                 return StatusCode(500, "Internal Server Error " + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetInvitesByInviter")]
         public async Task<IActionResult> GetInvitesByInviterId(string inviterId)
         {
             try
             {
-                var invites = await _inviteRepo.GetInvitesByInviterAsync(inviterId);
-                if (invites.Count() == 0) return Ok(new List<InviteDto>());
-              
+                if (!await _userRepo.UserExists(inviterId))
+                {
+                    return BadRequest("Inviter does not exists!");
+                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
 
-                var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
-                return Ok(invitesDto);
+                if (userId == inviterId || userTokenRole == "Admin")
+                {
+                    var invites = await _inviteRepo.GetInvitesByInviterAsync(inviterId);
+                    if (invites.Count() == 0) return Ok(new List<InviteDto>());
+             
+                    var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
+                    return Ok(invitesDto);   
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error " + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetInvitesByInvitee")]
         public async Task<IActionResult> GetInvitesByInviteeId(string inviteeId)
         {
             try
             {
-                var invites = await _inviteRepo.GetInvitesByInviteeAsync(inviteeId);
-                if (invites.Count() == 0) return Ok(new List<InviteDto>());
+                if (!await _userRepo.UserExists(inviteeId))
+                {
+                    return BadRequest("Inviter does not exists!");
+                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+
+                if (userId == inviteeId || userTokenRole == "Admin")
+                {
+                    var invites = await _inviteRepo.GetInvitesByInviteeAsync(inviteeId);
+                    if (invites.Count() == 0) return Ok(new List<InviteDto>());
 
 
-                var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
-                return Ok(invitesDto);
+                    var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
+                    return Ok(invitesDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
@@ -98,54 +121,85 @@ namespace backend.Controllers
 
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetInvitesByWorkspace")]
         public async Task<IActionResult> GetInvitesByWorkspace(int workspaceId)
         {
             try
             {
-                var invites = await _inviteRepo.GetInvitesByWorkspaceAsync(workspaceId);
-                if (invites.Count() == 0) return Ok(new List<InviteDto>());
-                
-
-                var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
-                return Ok(invitesDto);
+                if (!await _workspaceRepo.WorkspaceExists(workspaceId))
+                {
+                    return BadRequest("Workspace not found!");
+                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, workspaceId);
+                if (ownsWorkspace || userTokenRole == "Admin")
+                {
+                    var invites = await _inviteRepo.GetInvitesByWorkspaceAsync(workspaceId);
+                    if (invites.Count() == 0) return Ok(new List<InviteDto>());
+                    var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
+                    return Ok(invitesDto);    
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error " + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetPendingInvitesByInviter")]
-        public async Task<IActionResult> GetPendingInvitesByInviter(string userId)
+        public async Task<IActionResult> GetPendingInvitesByInviter(string inviterId)
         {
             try
             {
-                var invites = await _inviteRepo.GetPendingInvitesByInviter(userId);
-                if (invites.Count() == 0) return Ok(new List<InviteDto>());
-                var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
-                return Ok(invitesDto);
+                if (!await _userRepo.UserExists(inviterId))
+                {
+                    return BadRequest("User not found!");
+                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                if (userId == inviterId || userTokenRole == "Admin")
+                {
+                    var invites = await _inviteRepo.GetPendingInvitesByInviter(inviterId);
+                    if (invites.Count() == 0) return Ok(new List<InviteDto>());
+                    var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
+                    return Ok(invitesDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }  catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error " + e.Message);
             }
         }
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpGet("GetPendingInvitesByInvitee")]
-        public async Task<IActionResult> GetPendingInvitesByInvitee(string userId)
+        public async Task<IActionResult> GetPendingInvitesByInvitee(string inviteeId)
         {
             try
             {
-                var invites = await _inviteRepo.GetPendingInvitesByInvitee(userId);
-                if (invites.Count() == 0) return Ok(new List<InviteDto>());
-                var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
-                return Ok(invitesDto);
+                if (!await _userRepo.UserExists(inviteeId))
+                {
+                    return BadRequest("User not found!");
+                }
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+
+                if (userId == inviteeId || userTokenRole == "Admin")
+                {
+                    var invites = await _inviteRepo.GetPendingInvitesByInvitee(inviteeId);
+                    if (invites.Count() == 0) return Ok(new List<InviteDto>());
+                    var invitesDto = _mapper.Map<IEnumerable<InviteDtoOut>>(invites);
+                    return Ok(invitesDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }  catch (Exception e)
             {
                 return StatusCode(500, "Internal Server Error " + e.Message);
             }
         }
-
+        [Authorize(AuthenticationSchemes = "Bearer")]
         [HttpPost("Invite")]
         public async Task<IActionResult> Invite(InviteDto inviteDto)
         {
@@ -178,37 +232,53 @@ namespace backend.Controllers
                 return BadRequest("Invitee is already a member of the workspace.");
             }
 
-            
             try
             {
-                var inviteModel = _mapper.Map<Invite>(inviteDto);
-                await _inviteRepo.AddInviteAsync(inviteModel);
-                return CreatedAtAction(nameof(GetInviteById), new { id = inviteModel.InviteId },
-                    _mapper.Map<InviteDto>(inviteModel));
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                if (userId == inviteDto.InviterId || userTokenRole == "Admin")
+                {
+                    var inviteModel = _mapper.Map<Invite>(inviteDto);
+                    await _inviteRepo.AddInviteAsync(inviteModel);
+                    return CreatedAtAction(nameof(GetInviteById), new { id = inviteModel.InviteId },
+                        _mapper.Map<InviteDto>(inviteModel));
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
-                return StatusCode(500, "Internal Server Error!"+e.Message);
+                return StatusCode(500, "Internal Server Error!" + e.Message);
             }
 
         }
 
-        [HttpPut]
-        [Route("UpdateInvite")]
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPut("UpdateInvite")]
         public async Task<IActionResult> UpdateInvite(UpdateInviteDto updateDto)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
-            }
-
+            } 
             try
             {
-                var invite = await _inviteRepo.UpdateInviteStatusAsync(updateDto);
-                if (invite == null) return NotFound("Invite Not Found!");
+                var invite = await _inviteRepo.GetInviteByIdAsync(updateDto.InviteId);
+                if (invite == null) return BadRequest("Invite not found!");
+                
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                var inviterId = invite.InviterId;
+                var inviteeId = invite.InviteeId;
+                if (userId == inviteeId || userId == inviteeId || userTokenRole == "Admin")
+                {
 
-                var inviteDto = _mapper.Map<InviteDtoOut>(invite);
-                return Ok(inviteDto);
+                    var inviteUpdate = await _inviteRepo.UpdateInviteStatusAsync(updateDto);
+                    if (inviteUpdate == null) return NotFound("Invite Not Found!");
+
+                    var inviteDto = _mapper.Map<InviteDtoOut>(inviteUpdate);
+                    return Ok(inviteDto);
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
@@ -216,21 +286,33 @@ namespace backend.Controllers
             }
         }
 
-        [HttpDelete]
-        [Route("DeleteInviteById")]
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpDelete("DeleteInviteById")]
         public async Task<IActionResult> DeleteInviteById([FromQuery] InviteIdDto inviteIdDto)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-
+            
             try
             {
-                var inviteModel = await _inviteRepo.DeleteInviteAsync(inviteIdDto.InviteId);
-                if (inviteModel == null) return NotFound("Invite Not Found!");
+                var invite = await _inviteRepo.GetInviteByIdAsync(inviteIdDto.InviteId);
 
-                return Ok("Invite Deleted!");
+                if (invite == null) return BadRequest("Invite not found!");
+                
+                var inviterId = invite.InviterId;
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
+                var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+                if (userId == inviterId || userTokenRole == "Admin")
+                {
+
+                    var inviteModel = await _inviteRepo.DeleteInviteAsync(inviteIdDto.InviteId);
+                    if (inviteModel == null) return NotFound("Invite Not Found!");
+
+                    return Ok("Invite Deleted!");
+                }
+                return StatusCode(401, "You are not authorized!");
             }
             catch (Exception e)
             {
